Time: 14 ms (65.59%), Space: 45.5 MB (24.21%) - LeetHub

+++ 0347-top-k-frequent-elements/0347-top-k-frequent-elements.java
@@ -20,7 +20,7 @@ class Solution {
         int count =0;
         while(!pq.isEmpty())
         {
-            ans[count++] = pq.poll().getKey();
+            ans[count++] = pq.poll().getKey().intValue();
         }
         return ans; 
     }
